pyproject.toml
[project]
name = "cupt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.121.0",
    "greenlet>=3.2.4",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "python-telegram-bot>=22.5",
    "ruff>=0.14.3",
]
